Move createProvider inside Provider.js
