+++ src/index.js
@@ -1,5 +1,4 @@
-import Provider from './components/Provider'
-import createProvider from './components/createProvider'
+import Provider, { createProvider } from './components/Provider'
 import connectAdvanced from './components/connectAdvanced'
 import connect from './connect/connect'
